Move NicType to nic.py
